Allow $x^\prime$ syntax
And also $x^'$, although this used to be an error
in LaTeX.

This code needs improvement, but it works for now.

Closes #119

Note this changes the output of the test X200a;
we need to think about what the most desirable
outcomes here are. In facilitating the "^\prime"
syntax, we've now got slightly visually incompatible
output in the differences between $\prime$ and
$^\prime$ (technically this shouldn't happen, I
think, so perhaps this problem can be solved in LuaTeX
with the appropriate select of font dimensions).