remove growing-accents feature
I'm sorry to be removing tested code, but I still strongly believe that \hat and \tilde should never stretch if we can help it.

We don't want to add a feature if we can avoid it. Is there a good reason to want \hat to stretch? I think not, and therefore the feature isn't worth the additional complexity.

closes #115